webrtc distributed

+++ src/placeos-rest-api/controllers/webrtc/chat_manager.cr
@@ -2,24 +2,6 @@ require "http"
 require "./kick_reason"
 
 module PlaceOS::Api
-  class CallDetails
-    include JSON::Serializable
-
-    getter id : String
-    getter peers : Hash(String, HTTP::WebSocket)
-
-    @[JSON::Field(converter: Time::EpochConverter)]
-    getter created_at : Time
-
-    @[JSON::Field(converter: Time::EpochConverter)]
-    property updated_at : Time
-
-    def initialize(@id : String)
-      @peers = {} of String => HTTP::WebSocket
-      @updated_at = @created_at = Time.utc
-    end
-  end
-
   # use a manager so the we can free the request context objects
   class ChatManager
     Log = ::Log.for(self)
@@ -27,14 +9,25 @@ module PlaceOS::Api
     def initialize(@ice_config)
       # grab the existing `PlaceOS::Driver::Subscriptions` instance
       subscriber = PlaceOS::Api::WebSocket::Session.subscriptions.@subscriber
-      subscriber.channel("chat/forward_signal") do ||
-        #TODO
+      subscriber.channel("internal/chat/forward_signal") do |_, payload|
+        signal = SessionSignal.from_json(payload)
+        perform_forwarded_signal(signal)
+      end
+      subscriber.channel("internal/chat/kick_user") do |_, payload|
+        user_id, reason = Tuple(String, String).from_json(payload)
+        perform_kick_user(user_id, reason)
+      end
+      subscriber.channel("internal/chat/transfer_user") do |_, payload|
+        user_id, session_id, connection_details = Tuple(String, String?, String?).from_json(payload)
+        perform_transfer(user_id, session_id, connection_details)
       end
 
       spawn { ping_sockets }
     end
 
-    SESSIONS = ClusteredSessions.new
+    # =================================
+    # Websocket Ping / ensure connected
+    # =================================
 
     protected def ping_sockets
       loop do
@@ -64,9 +57,32 @@ module PlaceOS::Api
     rescue
     end
 
+    # =================================
+    # Various helpers
+    # =================================
+
     protected def redis_publish(path : String, payload)
       ::PlaceOS::Driver::RedisStorage.with_redis &.publish(path, payload.to_json)
     end
+
+    def create_new_call(signal) : CallDetails
+      calls[signal.session_id] = CallDetails.new(signal.session_id)
+    end
+
+    def send_signal(websocket, signal)
+      Log.trace { "Sending signal #{signal.type} to #{signal.session_id}" }
+      websocket.send(signal.to_json)
+    rescue
+      # we'll ignore websocket send failures, the user will be cleaned up
+    end
+
+    def member_list(session_id : String) : Array(String)
+      CallDetails::SESSIONS.user_list(session_id)
+    end
+
+    # =================================
+    # Connection management
+    # =================================
 
     # authority_id => config string
     private getter ice_config : Hash(String, String)
@@ -113,37 +129,28 @@ module PlaceOS::Api
     end
 
     def remove_from_call(connect_details : SessionSignal)
-      call = calls[connect_details.session_id]?
-      return unless call
+      session_id = connect_details.session_id
 
-      # inform the call peers that the user is gone
-      call.peers.delete connect_details.user_id
-      call.updated_at = Time.utc
-
-      connect_details.type = :leave
-      call.peers.each_value do |ws|
-        send_signal(ws, connect_details)
+      if call = calls[session_id]?
+        # inform the call peers that the user is gone
+        call.remove connect_details.user_id
+        # cleanup empty sessions
+        calls.delete(session_id) if call.peers.empty?
       end
 
-      # cleanup empty sessions
-      calls.delete(connect_details.session_id) if call.peers.empty?
-    end
-
-    def create_new_call(signal) : CallDetails
-      calls[signal.session_id] = CallDetails.new(signal.session_id)
-    end
-
-    def send_signal(websocket, signal)
-      Log.trace { "Sending signal #{signal.type} to #{signal.session_id}" }
-      websocket.send(signal.to_json)
-    rescue
-      # we'll ignore websocket send failures, the user will be cleaned up
+      # forward the leave signal to all the members of the call
+      connect_details.type = :leave
+      CallDetails::SESSIONS.user_list(session_id).each do |user_id|
+        connect_details.to_user = user_id
+        redis_publish("placeos/internal/chat/forward_signal", connect_details)
+      end
     end
 
     def on_join_signal(websocket, signal, auth_id) : Nil
       call = calls[signal.session_id]? || create_new_call(signal)
 
       # check the current user can join the call (prevent spoofing)
+      # TODO:: look into this now the service is clustered
       if existing_peer_ws = call.peers[signal.user_id]?
         if existing_user = sockets[existing_peer_ws]?
           if existing_user.place_user_id != signal.place_user_id
@@ -190,6 +197,10 @@ module PlaceOS::Api
       })
     end
 
+    # ================================
+    # Forward signal
+    # ================================
+
     def forward_signal(websocket, signal) : Nil
       if call = calls[signal.session_id]?
         # check the current user is in the call
@@ -207,23 +218,28 @@ module PlaceOS::Api
           return
         end
 
+        redis_publish("placeos/internal/chat/forward_signal", signal)
+      end
+    end
+
+    # all security checks have occured at this point, forward the message
+    # if the user is connected to this server
+    protected def perform_forwarded_signal(signal)
+      if call = calls[signal.session_id]?
         if to_user = call.peers[signal.to_user]?
           send_signal(to_user, signal)
         end
       end
     end
 
-    enum TransferResult
-      NoConnection
-      NoSession
-      SignalSent
-    end
+    # ================================
+    # Kick User / User exited
+    # ================================
 
     # the user has exited chat
     def end_call(user_id : String, auth_id : String)
       # find the users websocket
-      websocket = user_lookup[user_id]?
-      websocket.try(&.close)
+      redis_publish("placeos/internal/chat/kick_user", {user_id, "call ended"})
 
       # signal the user exited
       redis_publish("placeos/#{auth_id}/chat/user/exited", {
@@ -231,7 +247,17 @@ module PlaceOS::Api
       })
     end
 
-    def kick_user(user_id : String, session_id : String, details : KickReason)
+    def kick_user(auth_id : String, user_id : String, session_id : String, details : KickReason)
+      # find the users websocket
+      redis_publish("placeos/internal/chat/kick_user", {user_id, details.reason})
+
+      # TODO:: need the auth id here!
+      redis_publish("placeos/#{auth_id}/chat/user/exited", {
+        user_id: user_id,
+      })
+    end
+
+    def perform_kick_user(user_id, reason)
       # find the users websocket
       websocket = user_lookup[user_id]?
       return unless websocket
@@ -239,40 +265,45 @@ module PlaceOS::Api
       connect_details = sockets[websocket]?
       return unless connect_details
 
-      # remove the user from the current call
-      remove_from_call(connect_details)
-
-      redis_publish("placeos/#{connect_details.place_auth_id}/chat/user/exited", {
-        user_id: connect_details.user_id,
-      })
-
       # send the kicked user a leave signal
       send_signal(websocket, SessionSignal.new(
         id: "SIGNAL::#{Time.utc.to_unix_ms}+#{Random::Secure.hex(6)}",
         type: :leave,
-        session_id: session_id,
+        session_id: connect_details.session_id,
         user_id: user_id,
         to_user: user_id,
-        value: details.to_json
+        value: KickReason.new(reason).to_json
       ))
+
+      websocket.close
     end
 
-    def member_list(session_id : String) : Array(String)
-      if call = calls[session_id]?
-        call.peers.keys
-      else
-        [] of String
-      end
+    # ================================
+    # Transfer user
+    # ================================
+
+    enum TransferResult
+      NoConnection
+      NoSession # not currently used
+      SignalSent
     end
 
     # transfer a user to a new chat room
     def transfer(user_id : String, session_id : String? = nil, payload : String? = nil) : TransferResult
+      current_session_id = CallDetails::SESSIONS.lookup_session(user_id)
+      return TransferResult::NoConnection unless current_session_id
+
+      redis_publish("placeos/internal/chat/transfer_user", {user_id, session_id, payload})
+      TransferResult::SignalSent
+    end
+
+    def perform_transfer(user_id : String, session_id : String? = nil, payload : String? = nil)
       # find the users websocket
       websocket = user_lookup[user_id]?
-      return TransferResult::NoConnection unless websocket
+      return unless websocket
 
       connect_details = sockets[websocket]?
-      return TransferResult::NoSession unless connect_details
+      return unless connect_details
 
       # remove the user from the current call
       remove_from_call(connect_details) if session_id && session_id != connect_details.session_id
@@ -286,7 +317,6 @@ module PlaceOS::Api
         to_user: user_id,
         value: payload
       ))
-      TransferResult::SignalSent
     end
   end
 end
